feat: 增加数据源接口定时执行更新熟悉数据任务

+++ api-model/src/main/java/xyz/linyh/model/dscInterfaceColumn/entitys/DscInterfaceColumn.java
@@ -19,12 +19,12 @@ public class DscInterfaceColumn implements Serializable {
      * 
      */
     @TableId(type = IdType.AUTO)
-    private Integer id;
+    private Long id;
 
     /**
      * 接口id
      */
-    private Integer interfaceInfoId;
+    private Long interfaceInfoId;
 
     /**
      * schemaName
